requirements.txt
Flask==1.0.2
pyjwt==1.4.2
